Update formatting test to use Symfony\Console output for piping
